# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WikiLLM API
  description: |-
    HTTP API for [wiki-llm](https://gist.github.com/karpathy/442a6bf555914893e9891c11519de94f)
    folders. Designed to run alongside Obsidian, git, LiveSync, and other sync tools while
    serving multiple users and agents from the same markdown wiki.

    ## Project links

    - [GitHub repository](https://github.com/LukasParke/wikillm-api)
    - [Docker image](https://github.com/LukasParke/wikillm-api/pkgs/container/wikillm-api)
    - [Deployment guide](https://github.com/LukasParke/wikillm-api#deployment)
    - [Architecture docs](https://github.com/LukasParke/wikillm-api/blob/main/docs/ARCHITECTURE.md)

    ## Features

    - Self-hosted knowledge base service for humans and agents
    - Filesystem remains the source of truth; the index is a rebuildable cache
    - Atomic writes with optimistic concurrency control
    - Multi-source API key attribution with OKF actor convention
    - Hybrid retrieval: full-text + vector + recency decay, RRF fusion (K=60), optional rerank
    - LLM answering with citations (planner → executor → synthesis)
    - SSE and WebSocket live change feeds
    - Batch ingestion of sources and wiki pages
    - Connector framework (git, web, GitHub)
    - MCP server exposing LLM-free retrieval tools
    - Projects scoping with per-key RBAC (read/write/admin)
    - Google OKF v0.2 conformance: validation, trust tiers, provenance stamping
  version: 0.1.0
  contact:
    name: WikiLLM API
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key passed as a Bearer token. Keys are configured via the API_KEYS environment variable.
  schemas:
...
    Conflict:
      type: object
      required:
        - hash
        - content
      properties:
        hash:
          type: string
        content:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: conflict
        message:
          type: string
        current:
          $ref: "#/components/schemas/Conflict"
        existingHash:
          type: string
...
    DocumentEntry:
      type: object
      required:
        - rel_path
        - kind
        - hash
        - mtime
      properties:
        rel_path:
          type: string
          example: wiki/entities/OpenAI.md
        kind:
          type: string
          enum:
            - page
            - source
            - doc
          description: page (wiki markdown), source (raw bytes), doc (connector-managed)
        origin:
          type: string
          nullable: true
          description: Connector id or actor that produced the document
        title:
          type: string
          nullable: true
        okf_type:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          nullable: true
        stale_after:
          type: string
          format: date-time
          nullable: true
        trust:
          type: string
          nullable: true
          enum:
            - unverified
            - machine-confirmed
            - human-reviewed
        hash:
          type: string
          description: SHA-256 of the document content
        mtime:
          type: number
          description: File modification time in milliseconds
        updated_at:
          type: string
          format: date-time
          nullable: true
        updated_by:
          type: string
          nullable: true
    DocumentList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentEntry"
        nextCursor:
          type: string
          nullable: true
...
    BatchOperationResult:
      type: object
      required:
        - rel_path
        - success
      properties:
        rel_path:
          type: string
        success:
          type: boolean
        error:
          type: string
          nullable: true
        conflict:
          $ref: "#/components/schemas/Conflict"
...
    DeleteDocumentsInput:
      type: object
      required:
        - rel_paths
      properties:
        rel_paths:
          type: array
          maxItems: 1000
          items:
            type: string
    DeleteDocumentsResult:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/BatchOperationResult"
...
paths:
...
  /v1/documents:
    get:
      tags:
        - Documents
      summary: List every indexed document
      description: |
        General listing across wiki pages, raw sources, and connector docs. Supports
        ETag revalidation: send If-None-Match with the previously returned ETag to
        receive 304 when nothing changed.
      parameters:
        - name: kind
          in: query
          schema:
            type: string
            enum:
              - page
              - source
              - doc
        - name: origin
          in: query
          schema:
            type: string
          description: Filter by origin (connector id or actor)
        - name: folder
          in: query
          schema:
            type: string
          description: Restrict to a folder prefix
        - name: type
          in: query
          schema:
            type: string
          description: Comma-separated OKF type filter
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tag filter (all must match)
        - name: status
          in: query
          schema:
            type: string
          description: Comma-separated lifecycle status filter
        - name: trust
          in: query
          schema:
            type: string
            enum:
              - unverified
              - machine-confirmed
              - human-reviewed
          description: Minimum derived trust tier
        - name: fresh
          in: query
          schema:
            type: boolean
          description: When true, exclude documents past their stale_after date
        - name: project
          in: query
          schema:
            type: string
          description: Restrict to a project scope
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
          description: Path to start after for pagination
      responses:
        "200":
          description: Document list
          headers:
            ETag:
              schema:
                type: string
              description: Collection fingerprint for If-None-Match revalidation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
        "304":
          description: Collection unchanged since the supplied If-None-Match
...
  /v1/documents/{rel_path}/content:
    parameters:
      - name: rel_path
        in: path
        required: true
        schema:
          type: string
        description: Relative path of any indexed document
        example: wiki/entities/OpenAI.md

    get:
      tags:
        - Documents
      summary: Download a document's content
      description: |
        Dispatches on the indexed kind: pages return text/markdown bodies, sources
        return their original bytes with the stored Content-Type, and connector docs
        return the indexed body held in the store. Responses carry an ETag header.
      responses:
        "200":
          description: Document content
          headers:
            Content-Type:
              schema:
                type: string
            ETag:
              schema:
                type: string
          content:
            text/markdown:
              schema:
                type: string
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/documents/delete:
    post:
      tags:
        - Documents
      summary: Bulk delete documents
      description: |
        Deletes files by relative path and returns one result per operation.
        Connector-managed docs fail with error "connector_managed" — delete them
        via their connector instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDocumentsInput"
      responses:
        "200":
          description: Per-operation deletion results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteDocumentsResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/bundle/export:
    get:
      tags:
        - Bundle
      summary: Export the bundle
      description: |
        Streams the knowledge base as a .tar.gz archive. Without filters the whole
        bundle is exported (X-Exported-Files: all); with filters only matching files
        are included and X-Exported-Files reports their count. Returns 404 when the
        filter set matches nothing.
      parameters:
        - name: prefix
          in: query
          schema:
            type: string
          description: Folder prefix filter
        - name: kind
          in: query
          schema:
            type: string
            enum:
              - page
              - source
          description: Only include this kind (connector docs have no backing file)
        - name: origin
          in: query
          schema:
            type: string
          description: Only files produced by this origin (connector id or actor)
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: ISO watermark; intersects candidates with the changes ledger
        - name: project
          in: query
          schema:
            type: string
          description: Restrict to a project scope
      responses:
        "200":
          description: Bundle tarball
          headers:
            X-Exported-Files:
              schema:
                type: string
              description: Number of exported files, or "all" for unfiltered exports
          content:
            application/gzip:
              schema:
                type: string
                format: binary
        "404":
          description: No files match the export filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/bundle/import:
    post:
      tags:
        - Bundle
      summary: Import a bundle
      description: |
        Imports a previously exported .tar.gz bundle (admin only). Without force,
        conflicting paths return 409 with the conflict list.
      parameters:
        - name: force
          in: query
          schema:
            type: boolean
            default: false
          description: Overwrite conflicting paths
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Bundle imported
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                properties:
                  success:
                    type: boolean
        "409":
          description: Conflicting paths without force
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...